+++ viewmodel/src/commonMain/kotlin/org/rhasspy/mobile/viewmodel/configuration/connections/http/list/HttpConnectionListConfigurationUiEvent.kt
@@ -6,8 +6,8 @@ sealed interface HttpConnectionListConfigurationUiEvent {
 
         data object BackClick : Action
         data object AddClick : Action
-        data class ItemClick(val id: Int) : Action
-        data class ItemDeleteClick(val id: Int) : Action
+        data class ItemClick(val id: Long) : Action
+        data class ItemDeleteClick(val id: Long) : Action
 
     }
